Rename utils to testing_utils.
